Domain layer model names are simplified rather than having small models with complex and quite same type of names its rather good to have a single large model. Field names are too simplified in Dto ,model object The search functionality is just implemented in the data module thus repository and api. The forecast route is cleaned up, some Ui problems still exits but need to find those A simple about page is created for information about the api being used. A graph is added to visualize the WeatherForecast From build.gradle unwanted dependencies are removed : systemUiController and swipe to refresh. For storing the locations for the cities we will be using preference dataStore for that a Facade is create.

+++ app/src/main/java/com/eva/androidweatherapp/data/remote/dto/common/SearchedLocationDto.kt
@@ -8,7 +8,7 @@ data class SearchedLocationDto(
     @SerialName("country") val country: String,
     @SerialName("lat") val latitude: Double,
     @SerialName("localtime") val time: String,
-    @SerialName("localtime_epoch") val epoch: Int,
+    @SerialName("localtime_epoch") val epoch: Long,
     @SerialName("lon") val longitude: Double,
     @SerialName("name") val name: String,
     @SerialName("region") val region: String,
